Isolate standard precompiles from benchmarks to reuse in tests
