+++ src/benches/eth_standard_precompiles.rs
@@ -1,10 +1,8 @@
 use crate::prelude::U256;
-use crate::transaction::EthTransaction;
 use criterion::{BatchSize, BenchmarkId, Criterion};
 use secp256k1::SecretKey;
-use std::path::{Path, PathBuf};
 
-use crate::test_utils::solidity;
+use crate::test_utils::standard_precompiles::{PrecompilesConstructor, PrecompilesContract};
 use crate::test_utils::{address_from_secret_key, deploy_evm, sign_transaction, SUBMIT};
 
 const INITIAL_BALANCE: u64 = 1000;
@@ -77,87 +75,4 @@ pub(crate) fn eth_standard_precompiles_benchmark(c: &mut Criterion) {
     }
 
     group.finish();
-}
-
-struct PrecompilesConstructor(solidity::ContractConstructor);
-
-struct PrecompilesContract(solidity::DeployedContract);
-
-impl From<PrecompilesConstructor> for solidity::ContractConstructor {
-    fn from(c: PrecompilesConstructor) -> Self {
-        c.0
-    }
-}
-
-impl PrecompilesConstructor {
-    fn load() -> Self {
-        Self(solidity::ContractConstructor::compile_from_source(
-            Self::sources_root(),
-            Self::solidity_artifacts_path(),
-            "StandardPrecompiles.sol",
-            "StandardPrecompiles",
-        ))
-    }
-
-    fn deploy(&self, nonce: U256) -> EthTransaction {
-        let data = self
-            .0
-            .abi
-            .constructor()
-            .unwrap()
-            .encode_input(self.0.code.clone(), &[])
-            .unwrap();
-        EthTransaction {
-            nonce,
-            gas_price: Default::default(),
-            gas: Default::default(),
-            to: None,
-            value: Default::default(),
-            data,
-        }
-    }
-
-    fn solidity_artifacts_path() -> PathBuf {
-        Path::new("target").join("solidity_build")
-    }
-
-    fn sources_root() -> PathBuf {
-        Path::new("src").join("benches").join("res")
-    }
-}
-
-impl PrecompilesContract {
-    fn call_method(&self, method_name: &str, nonce: U256) -> EthTransaction {
-        let data = self
-            .0
-            .abi
-            .function(method_name)
-            .unwrap()
-            .encode_input(&[])
-            .unwrap();
-        EthTransaction {
-            nonce,
-            gas_price: Default::default(),
-            gas: Default::default(),
-            to: Some(self.0.address),
-            value: Default::default(),
-            data,
-        }
-    }
-
-    fn all_method_names() -> &'static [&'static str] {
-        &[
-            "test_ecrecover",
-            "test_sha256",
-            "test_ripemd160",
-            "test_identity",
-            "test_modexp",
-            "test_ecadd",
-            "test_ecmul",
-            // TODO(#46): ecpair uses up all the gas (by itself) for some reason, need to look into this.
-            // "test_ecpair",
-            "test_blake2f",
-            "test_all",
-        ]
-    }
 }
